app-permission-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
pandas==2.2.3
scikit-learn==1.7.2
joblib==1.4.2
numpy==2.2.1
python-dotenv==1.0.1
pymongo==4.10.1
dnspython==2.8.0
requests==2.32.3
